+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -258,6 +258,12 @@ New checks
   Detects C++ code where a reference variable is used to extend the lifetime
   of a temporary object that has just been constructed.
 
+- New :doc:`readability-use-explicit-namespaces
+  <clang-tidy/checks/readability/use-explicit-namespaces>` check.
+
+  Detect and fix references to members of namespaces where the namespace is
+  not explicity specified in the reference.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^